+++ Dockerfile
@@ -29,7 +29,7 @@ WORKDIR /app
 COPY go.mod go.sum ./
 RUN go mod download
 
-COPY ./cmd/ ./cmd/
+COPY main.go ./
 COPY ./internal/ ./internal/
 COPY ./ebpf/ ./ebpf/
 
@@ -38,7 +38,7 @@ RUN go generate ./internal/redirect
 # CGO_ENABLED=0 to build a statically-linked binary
 # -ldflags '-w -s' to strip debugging information for smaller size
 RUN CGO_ENABLED=0 GOOS=linux go build -ldflags="-w -s" -o gke-metadata-server \
-    github.com/matheuscscp/gke-metadata-server/cmd
+    github.com/matheuscscp/gke-metadata-server
 
 FROM alpine:3.21
