+++ Dockerfile.test
@@ -29,7 +29,7 @@ WORKDIR /app
 COPY go.mod go.sum ./
 RUN go mod download
 
-COPY ./cmd/ ./cmd/
+COPY main.go ./
 COPY ./internal/ ./internal/
 COPY ./ebpf/ ./ebpf/
